requirements.txt
asgiref==3.7.2
Django==4.2.2
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-phone-field==1.8.1
djangorestframework==3.14.0
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
Pillow==9.5.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
